requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
osmnx==1.9.1
networkx==3.2.1
geopy==2.4.1
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
